Refactor: 修正 SwiftLint 錯誤並重構程式碼
本次提交主要解決了  發現的多個問題，並對相關程式碼進行了重構，以提高可讀性和可維護性。

主要變更：
- **SideBySideDiffHeaderView.swift**: 將  拆分為  和 ，以解決  的問題。
- **IntegratedRepositoryFileStatusTab.swift**: 將  拆分為  和 ，以解決  的問題。
- **DiffLineView.swift**: 將  拆分為多個輔助視圖，以解決  的問題，並移除了多餘的空白。
- **UnifiedDiffParser.swift**: 修正了 、、 和  的問題。
