+++ Sources/UIKit/Components/FileStatusListView.swift
@@ -8,46 +8,6 @@
 import GitCore
 import SwiftUI
 
-// MARK: - Supporting Types
-
-public enum FileStatusFilter: String, CaseIterable, Identifiable {
-    case all = "All Files"
-    case staged = "Staged Only"
-    case unstaged = "Modified Only"
-    case untracked = "Untracked Only"
-    case conflicted = "Conflicted Only"
-
-    public var id: String { rawValue }
-    public var systemImage: String {
-        switch self {
-        case .all: "doc.on.doc"
-        case .staged: "checkmark.circle"
-        case .unstaged: "pencil.circle"
-        case .untracked: "plus.circle"
-        case .conflicted: "exclamationmark.triangle"
-        }
-    }
-}
-
-public enum FileStatusGrouping: String, CaseIterable, Identifiable {
-    case none = "No Grouping"
-    case status = "Group by Status"
-    case directory = "Group by Directory"
-
-    public var id: String { rawValue }
-}
-
-public struct FileStatusGroup: Identifiable {
-    public let id = UUID()
-    public let title: String?
-    public let files: [GitStatusEntry]
-
-    public init(title: String?, files: [GitStatusEntry]) {
-        self.title = title
-        self.files = files
-    }
-}
-
 // MARK: - FileStatusListView Internal State
 
 @MainActor
@@ -194,7 +154,7 @@ public struct FileStatusListView: View {
             .navigationTitle("File Status")
             .toolbar {
                 ToolbarItem(placement: .primaryAction) {
-                    refreshButton
+                    FileStatusListUtilityViews(repository: repository, stagingViewModel: stagingViewModel).refreshButton
                 }
             }
             .focusable()
@@ -203,314 +163,25 @@ public struct FileStatusListView: View {
     private var mainView: some View {
         VStack(spacing: 0) {
             // Header with controls
-            headerView
+            FileStatusListHeaderView(
+                state: state,
+                stagingViewModel: stagingViewModel,
+                fileStatuses: fileStatuses
+            )
 
             Divider()
 
             // Content area
-            contentView
-
-            // Message displays
-            messageDisplays
-        }
-    }
-
-    private var messageDisplays: some View {
-        VStack(spacing: 0) {
-            // Error message display
-            if let errorMessage = stagingViewModel.errorMessage {
-                errorBanner(message: errorMessage)
-            }
-
-            // Success message display
-            if let successMessage = stagingViewModel.lastOperationSuccessMessage {
-                successBanner(message: successMessage)
-            }
-        }
-    }
-
-    // MARK: - Header View
-
-    private var headerView: some View {
-        VStack(spacing: 12) {
-            searchAndFilterControls
-            groupingAndSelectionControls
-            statusSummary
-        }
-        .padding()
-        .background(Color(NSColor.windowBackgroundColor))
-    }
-
-    private var searchAndFilterControls: some View {
-        HStack {
-            // Search field
-            HStack {
-                Image(systemName: "magnifyingglass")
-                    .foregroundColor(.secondary)
-                TextField("Search files...", text: $state.searchText)
-                    .textFieldStyle(.plain)
-            }
-            .padding(.horizontal, 8)
-            .padding(.vertical, 6)
-            .background(Color(NSColor.controlBackgroundColor))
-            .cornerRadius(6)
-
-            // Filter picker
-            Picker("Filter", selection: $state.selectedFilter) {
-                ForEach(FileStatusFilter.allCases) { filter in
-                    Label(filter.rawValue, systemImage: filter.systemImage)
-                        .tag(filter)
-                }
-            }
-            .pickerStyle(.menu)
-            .frame(width: 140)
-        }
-    }
-
-    private var groupingAndSelectionControls: some View {
-        HStack {
-            // Grouping picker
-            Picker("Group by", selection: $state.selectedGrouping) {
-                ForEach(FileStatusGrouping.allCases) { grouping in
-                    Text(grouping.rawValue)
-                        .tag(grouping)
-                }
-            }
-            .pickerStyle(.segmented)
-
-            Spacer()
-
-            // Selection controls
-            if state.hasSelection {
-                selectionControls
-            }
-
-            // Global staging actions
-            if !state.hasSelection {
-                globalStagingControls
-            }
-        }
-    }
-
-    private var selectionControls: some View {
-        HStack(spacing: 8) {
-            Text("\(state.selectionCount) selected")
-                .font(.caption)
-                .foregroundColor(.secondary)
-
-            Button("Select All") { state.selectAll(from: fileStatuses) }.font(.caption)
-
-            Button("Deselect All") { state.deselectAll() }.font(.caption)
-
-            // Staging operations for selected files
-            if state.hasSelection {
-                Divider()
-                    .frame(height: 16)
-
-                stagingControlsForSelection
-            }
-        }
-    }
-
-    private var stagingControlsForSelection: some View {
-        HStack(spacing: 6) {
-            // Stage selected files
-            Button(action: {
-                Task {
-                    await stagingViewModel.stageSelectedFiles(state.selectedFiles)
-                    onStagingChanged?()
-                }
-            }) {
-                Label("Stage", systemImage: "plus.circle.fill")
-                    .font(.caption)
-            }
-            .disabled(stagingViewModel.isLoading)
-            .controlSize(.small)
-
-            // Unstage selected files
-            Button(action: {
-                Task {
-                    await stagingViewModel.unstageSelectedFiles(state.selectedFiles)
-                    onStagingChanged?()
-                }
-            }) {
-                Label("Unstage", systemImage: "minus.circle.fill")
-                    .font(.caption)
-            }
-            .disabled(stagingViewModel.isLoading)
-            .controlSize(.small)
-
-            // Toggle staging for selected files
-            Button(action: {
-                Task {
-                    await stagingViewModel.toggleSelectedFilesStaging(state.selectedFiles)
-                    onStagingChanged?()
-                }
-            }) {
-                Label("Toggle", systemImage: "arrow.up.arrow.down.circle")
-                    .font(.caption)
-            }
-            .disabled(stagingViewModel.isLoading)
-            .controlSize(.small)
-        }
-    }
-
-    private var globalStagingControls: some View {
-        HStack(spacing: 8) {
-            stagingButton("Stage All", icon: "plus.circle.fill", disabled: !stagingViewModel.hasChangesToStage) {
-                await stagingViewModel.stageAllFiles()
-                onStagingChanged?()
-            }
-            stagingButton("Unstage All", icon: "minus.circle.fill", disabled: !stagingViewModel.hasStagedChanges) {
-                await stagingViewModel.unstageAllFiles()
-                onStagingChanged?()
-            }
-            stagingButton("Stage Modified", icon: "pencil.circle.fill") {
-                await stagingViewModel.stageModifiedFiles()
-                onStagingChanged?()
-            }
-            stagingButton("Stage Untracked", icon: "plus.circle.fill") {
-                await stagingViewModel.stageUntrackedFiles()
-                onStagingChanged?()
-            }
-        }
-    }
-
-    private func stagingButton(
-        _ title: String,
-        icon: String,
-        disabled: Bool = false,
-        action: @escaping () async -> Void
-    ) -> some View {
-        Button(action: { Task { await action() } }) {
-            Label(title, systemImage: icon).font(.caption)
-        }
-        .disabled(stagingViewModel.isLoading || disabled)
-        .controlSize(.small)
-    }
-
-    private var statusSummary: some View {
-        HStack {
-            let filtered = state.filteredFileStatuses(from: fileStatuses)
-            let counts = (
-                total: filtered.count,
-                staged: filtered.filter(\.isStaged).count,
-                modified: filtered.filter { $0.hasWorkingDirectoryChanges && !$0.isUntracked }.count,
-                untracked: filtered.filter(\.isUntracked).count,
-                conflicted: filtered.filter(\.hasConflicts).count
+            FileStatusListContentView(
+                state: state,
+                fileStatuses: fileStatuses,
+                onViewDiff: onViewDiff,
+                onStagingChanged: onStagingChanged,
+                stagingViewModel: stagingViewModel
             )
 
-            summaryBadge(count: counts.total, label: "Total", color: .primary)
-            if counts.staged > 0 { summaryBadge(count: counts.staged, label: "Staged", color: .green) }
-            if counts.modified > 0 { summaryBadge(count: counts.modified, label: "Modified", color: .orange) }
-            if counts.untracked > 0 { summaryBadge(count: counts.untracked, label: "Untracked", color: .gray) }
-            if counts.conflicted > 0 { summaryBadge(count: counts.conflicted, label: "Conflicts", color: .red) }
-            Spacer()
+            // Message displays
+            FileStatusListMessageView(stagingViewModel: stagingViewModel)
         }
-    }
-
-    private func summaryBadge(count: Int, label: String, color: Color) -> some View {
-        HStack(spacing: 2) {
-            Text("\(count)").fontWeight(.medium)
-            Text(label)
-        }
-        .font(.caption).foregroundColor(color)
-        .padding(.horizontal, 8).padding(.vertical, 2)
-        .background(color.opacity(0.1)).cornerRadius(4)
-    }
-
-    private var contentView: some View {
-        Group {
-            if state.filteredFileStatuses(from: fileStatuses).isEmpty {
-                emptyStateView
-            } else {
-                fileListView
-            }
-        }
-    }
-
-    private var emptyStateView: some View {
-        VStack(spacing: 16) {
-            Image(systemName: "checkmark.circle").font(.largeTitle).foregroundColor(.green)
-            Text("No Files Found").font(.headline)
-            Text(emptyStateMessage).foregroundColor(.secondary).multilineTextAlignment(.center)
-        }
-        .frame(maxWidth: .infinity, maxHeight: .infinity).padding()
-    }
-
-    private var emptyStateMessage: String {
-        if !state.searchText.isEmpty { return "No files match your search criteria." }
-        switch state.selectedFilter {
-        case .all: return "Working directory is clean. No changes detected."
-        case .staged: return "No files are currently staged for commit."
-        case .unstaged: return "No modified files in working directory."
-        case .untracked: return "No untracked files found."
-        case .conflicted: return "No conflicted files found."
-        }
-    }
-
-    private var fileListView: some View {
-        ScrollView {
-            LazyVStack(spacing: 8) {
-                ForEach(state.groupedFileStatuses(from: fileStatuses), content: groupView)
-            }.padding()
-        }
-    }
-
-    private func groupView(_ group: FileStatusGroup) -> some View {
-        VStack(alignment: .leading, spacing: 8) {
-            if let title = group.title {
-                HStack {
-                    Text(title).font(.headline).foregroundColor(.primary)
-                    Spacer()
-                    Text("\(group.files.count) files").font(.caption).foregroundColor(.secondary)
-                }.padding(.horizontal, 4)
-            }
-            ForEach(group.files) { fileStatus in
-                FileStatusItemView(
-                    fileStatus: fileStatus,
-                    isSelected: state.selectedFiles.contains(fileStatus.filePath),
-                    onSelectionToggle: { state.toggleFileSelection(fileStatus.filePath) },
-                    onStageToggle: {
-                        Task {
-                            await stagingViewModel.toggleFileStaging(fileStatus.filePath)
-                            onStagingChanged?()
-                        }
-                    },
-                    onViewDiff: onViewDiff.map { callback in
-                        { callback(fileStatus.filePath) }
-                    }
-                )
-            }
-        }
-    }
-
-    private var refreshButton: some View {
-        Button(action: {
-            Task {
-                await repository.refreshStatus()
-                await stagingViewModel.refreshStagingStatus()
-            }
-        }) { Image(systemName: "arrow.clockwise") }.help("Refresh file status")
-    }
-
-    private func errorBanner(message: String) -> some View {
-        banner(message: message, color: .red, icon: "exclamationmark.triangle.fill") { stagingViewModel.clearError() }
-    }
-
-    private func successBanner(message: String) -> some View {
-        banner(message: message, color: .green, icon: "checkmark.circle.fill") { stagingViewModel.clearLastResult() }
-    }
-
-    private func banner(message: String, color: Color, icon: String, onDismiss: @escaping () -> Void) -> some View {
-        HStack {
-            Image(systemName: icon).foregroundColor(color)
-            Text(message).font(.caption).foregroundColor(color)
-            Spacer()
-            Button("Dismiss", action: onDismiss).font(.caption).foregroundColor(color)
-        }
-        .padding(.horizontal, 16).padding(.vertical, 8)
-        .background(color.opacity(0.1)).cornerRadius(6)
-        .padding(.horizontal, 16).padding(.bottom, 8)
     }
 }
